\name{dim-methods}
\docType{methods}
\alias{dim-methods}
\alias{dim}
\alias{dim,DiscreteMVDistribution-method}

\title{ Methods for Function dim in Package `distrEx' }
\description{dim-methods}
\section{Methods}{\describe{
    \item{dim}{\code{signature(object = "DiscreteMVDistribution")}: returns the dimension of the distribution }
}}
\seealso{\code{\link[distr]{dim-methods}},\cr
 \code{\link[base]{dim}}}
\keyword{methods}
\concept{dimension}
\concept{space}
\concept{accessor function}
