\name{distrExMOVED}
\alias{distrExMOVED}
\alias{MOVEDING}
\title{Moved functionality from package "distrEx"}
\description{Provides information on moved  of
functionality from package \pkg{distrEx}. }
\usage{
distrExMOVED(library = NULL)
}
\arguments{
  \item{library}{a character vector with path names of R libraries, or \code{NULL}.
                 The default value of \code{NULL} corresponds to all libraries
                 currently known.  If the default is used, the loaded packages
                 are searched before the libraries}
    }
\value{no value is returned
}

\author{Peter Ruckdeschel \email{peter.ruckdeschel@uni-oldenburg.de}}
\examples{
distrExMOVED()
}
\keyword{programming}
\keyword{distribution}
\keyword{documentation}
\concept{moved functionality}
\concept{info file}
