\name{Naturals}
\alias{Naturals}

\title{Generating function for Naturals-class}
\description{
  Generates an object of class \code{"Naturals"}.
}
\usage{Naturals()}
%\arguments{}
%\details{}
\value{Object of class \code{"Naturals"}}
%\references{}
\author{Matthias Kohl \email{Matthias.Kohl@stamats.de}}
%\note{}
\seealso{\code{\link[distr]{Naturals-class}}}
\examples{Naturals()

## The function is currently defined as
function(){ new("Naturals") }
}
\concept{naturals}
\keyword{}
