\name{Reals}
\alias{Reals}

\title{Generating function for Reals-class}
\description{
  Generates an object of class \code{"Reals"}.
}
\usage{Reals()}
%\arguments{}
%\details{}
\value{Object of class \code{"Reals"}}
%\references{}
\author{Matthias Kohl \email{Matthias.Kohl@stamats.de}}
%\note{}
\seealso{\code{\link[distr]{Reals-class}}}
\examples{Reals()

## The function is currently defined as
function(){ new("Reals") }
}
\concept{reals}
\keyword{}
