\name{plot-methods}
\docType{methods}
\alias{plot}
\alias{plot-methods}
\alias{plot,UnivariateCondDistribution,missing-method}
\alias{plot,MultivariateDistribution,missing-method}
\title{ Methods for Function plot in Package `distrEx' }
\description{plot-methods}
\usage{plot(x, y, ...)
\S4method{plot}{UnivariateCondDistribution,missing}(x, y, ...)
\S4method{plot}{MultivariateDistribution,missing}(x, y, ...)
}
\arguments{
  \item{x}{object of class 
            \code{"UnivariateCondDistribution"} or class
            \code{"MultivariateDistribution"}: 
            distribution(s) which should be plotted }
  \item{y}{ missing }
  \item{\dots}{addtional arguments }
}
\details{
upto now only warnings are issued that the corresponding
method is not yet implemented;
}


